restructured examples component
